.github/scripts/requirements.txt
hvac==1.0.2
PyYAML==6.0
